Move handler callback protocols -> `kopf.reactor.callbacks`
They contain a lot of copy-pasting, so it is better to keep
them away from the meaningful code (handlers & registries).

As they are more protocols are going to be added, separating
them seems a good refactoring move to rebalance the modules.

Also, they are similar to `kopf.reactor.lifecycles`, as they
declare the invokable function signatures the same way.