Switch examples from `kubernetes` to `pykube-ng`
The examples are used as the e2e test. With the warnings converted
to errors, they fails quite often (randomly) due to SSL sockets
not closed properly (ResourceWarnings), noticed within the executor's
threads (which are used only for the sync-handlers).

Switching to `pykube-ng` and closing the request's session explicitly
(instead of on garbage collection) seems to fix it. So as not having
the session0owning object module-scoped.