Move handler field types to handlers
Once the handlers were moved away into its own separate module,
it makes no sense to keep the types of some of the handlers' fields
in other modules (such as causation).

Instead, it should be all together with handlers — and so we can move
the whole thing to the `kopf.structs` package eventually (next commit).
Because all of them are structs after all, not the reactor's processes.