Move in-memory handler execution back to "activities only"
Originally, the hope was to reuse them in daemons & timers. But this
breaks the semantics of the handlers: patches should be applied
on every exit, not only on successes; multiple restarts should
be counted & limited and distributed with backoffs; extra daemon
restarts are not needed at all if regular retrying works; etc.