Stop mocking the real sleep when it is switched to the sleeping engine
The real sleep is needed in some cases to give control to asyncio's
event loop (`sleep(0)`), or to debug the tests. When it is mocked,
it is difficult to do anything.

All the code uses the sleeping engine now, so there is no need
to check for the actual `asyncio.sleep()` calls.