Reuse pre-retrieved/pre-built old/new essences in causation & last-seen
Previously, the essence was regenerated in few places independently.
While it is okay in most cases, since the body/essence is never changed,
it is better to be safe by retrieving/generating the essences only once
in the beginning of a handling cycle, and then reusing the essences
in decision making. The resulting effect will be the same.

This will also save a little bit of CPU and RAM.