Run daemons only as long as the resource matches the filtering criteria
When the resource changes so that it does not match the filters,
stop the daemon.

When the resource changes so that it starts matching the filters,
start it (either for the first time or again if previously stopped).

This also covers a case when the criteria themselves are changed.
But it is only applied on the next resource watch-event (when any
activity is at all triggered).

Semantically, the daemon's filters define when the resource should be
accompanied on a continuous timeline, not only when it should be spawned
initially: i.e., not assuming both criteria and resource as constant.