Prevent premature removal of finalizers for multi-step deletion handlers
Technically: The finalizers removal criteria have changed from
"when done && reason == DELETE" to "when no delays returned"
(see 7adba026d949f9fe8d3b1420e296161b5eb207c3).

But the delays were not returned for multi-step deletion (2+ handlers),
so the finalizers were removed immediately after the 1st on-deletion
handler, and the 2nd and further on-deletion handlers were not invoked.

Now, with a delay of 0 for not yet finished (but not delayed) handlers,
it will not be considered as "done", which will prevent the deletion
of finalizers until really all handlers are finished either way.

See #356