Get rid of mocks for causes in the tests of registries
Causes need to be checked with `isinstance()` in the code base, to properly
pass the type checking: `isinstance()` adjusts the inferred type for operations
that follow the check. An example is handler-matching functions that check for
causes to be resource-changing causes, and using `.diff` only after that
(`.diff` is absent in all other causes).

However, this does not work with mocks in tests: all tests do fail, because
`Mock` and `MagicMock` classes do not pass the `isinstance()` criterion.

Besides that, mocks are just a terrible way of testing: first of all, they leak
the abstractions of the mocked objects by knowing which fields they contain,
which properties they implement and how (and can mimic them wrong); second,
they prevent tests to be type-checked both by type-checkers or by IDEs.