Sort imports with isort
Same-source "from"-imports are compacted into one statement,
same as PyCharm tries to auto-format them when refactoring.

This is now done only once manually, but later can become the automated
code linting step.

Note that the examples consider `kopf` as a third-party library --
the same as it would be with the real Kopf-based operators.