Remove the influence of watcher-workers exit polling, use async primitives
Previously, the 1/100th of the exit timeout could be a significant part
of overall timing of the watchers; especially in the unit-tests, where
the exit timeout is artificially set high enough (e.g. 100 seconds).

Now, asynchronous primitives are used to signal that the streams/jobs
have changed, and it is worth re-checking the streams depletion, and
exit the watcher if all the streams are depleted.

One special edge-case to consider: the last worker's signal will finish
the depletion waiter due to empty streams (the predicate becomes true).
However, when the signaller is notified, the worker job itself is still running.
It is assumed that the job finishes immediately after the notification,
specifically within the scheduler's `close_timeout` (0.1s), so that the watcher
can close the scheduler without any issues or leftovers.

This change allows the watcher-workers-queueing tests to be adjusted for proper
timing, so that we ignore the overhead introduced by polling the scheduler at
regular time intervals.