Separate the concerns of the tests: one test — one reason of failure
As a matter of good test design, a test should fail for one and only one reason.
Multiple reasons of failure should be done as multiple tests. In this case,
test that the original body is modified while storing the data into a patch,
but in a separate test with this exact reason of failure. Remove this reason
of failure from other tests that check for validity of the values stored.