Deprecate all internal registries for public use
Only leave `OperatorRegistry`, which is used as an argument to `registry=`
option of the handlers, and is the only one supposed to be used by the users
when they construct a custom operator. All other registries are implementation
details, and can change over time. Their interfaces are not part of the DSL/API.