Restore and purge all known handlers' states, not only the current ones
The current ones can vary, depending on the filters. For example, they can
exclude themselves from the following handling cycles. But the state remains
persisted until the full cause processing ends, and must be purged too.

For this, we have to try to restore all known handlers, not only those
selected at the moment. And we restore the persisted states for all of those
that have something persisted at all, even if those handlers are currently
not in the scope or do not match the filters. This ensures that nothing is lost.