Properly detect custom mappings (patches, raw bodies) in diffs
This should improve how different types of mappings are compared to each other,
ignoring their actual types: e.g. when a patch is compared with a raw body for
inconsistencies detection, the actual classes do not matter.

All non-equal non-mappings will be detected as a change of the whole value.