Drop deprecated "legacy peering mode" (KopfPeering, but cluster-wide)
The legacy peering mode is now 1.5 years since it was deprecated in the early stages. There is no need to support it anymore, as barely anyone still use it in their clusters (taking into account that peering per se it not used often at all). But the legacy mode complicates the coming features with dynamic cluster scanning. So, the time has come to drop it — without releasing the major version of the framework.