Substitute K8s API errors with our own classes
`aiohttp` is currently the underlying client library, and it raises its response errors. To enrich them with K8s's own explanations, the errors were extended by inheriting from the client library in #558. But anyway, this violates the principle that the API/HTTP client library is isolated to `kopf.clients`, and no single piece of it leaks outside to other packages (for replaceability).

Instead, introduce our own hierarchy of exceptions for K8s API errors, which are completely independent of the underlying library. The only connection to the original exceptions will be error chaining (`raise ... from ...`).

This allows us to analyse the API errors in other packages without coupling to the underlying library, while keeps the goal of readable stack traces and error explanations in the logs.