Ignore IPv6 ARPA hostnames for peering identities
With IPv6, the peering operators often got the identities like `1.0.0.0...0.0.ip6.arpa`, which does not make it easier to identify the pod where the operator is running. This is because of how `socket.getfqdn()` works — it does not exclude IPv6 pseudo-hostnames properly.

With this fix, the operators' identities in IPv6 environments will return back to normal pod names.