Fix CLI commands for peering
`kopf freeze` & `kopf resume` did not have the credentials vault and were broken all this time (they are used mostly for debugging the peering engine rather than for any real-world scenarios).

Fix them by providing the vault and the authenticating process which populates the vault.

The quality of the solution is irrelevant at the moment, as it is going to be redesigned soon into an arbitrary command wrapped into the operator's supporting tasks — but this requires the redesigned peering engine, which is too much for this tiny fix.