Deprecate any usage of internal registries
Internal registries are supposed for internal use only. They change the protocols and signatures all the time. Backward compatibility will not be ensured in the future.

The only usable registries are `OperatorRegistry` and `SmartOperatorRegistry`, and only to the extent that they can be created and used in the `registry=` kwarg to `@kopf.on...` decorators and `operator()` coroutine and related functions. All public fields and methods of these classes — selecting the handlers or extra-fields, adding new handlers, deleting the handlers, etc — are now deprecated.

It requires too much effort to maintain the internal registries backward compatible despite they are neither a primary nor a secondary feature of the framework — more like a side-effect of how there are (were) implemented.