Split resources into actual resources & selectors
In the next commits, resource selectors will be made flexible to contain any names and be used in the handler declarations, while the exact resource references will be discovered from the cluster at runtime. One selector (in a handler) could match multiple actual resources (in the cluster).

But the first atomic step to this is the separation of the types, and the usage of them in the code accordingly.

This split triggers the whole cascade of changes in how the registries are used: they are not structured as dicts with resources as keys anymore; instead, the registries are now flat, and the handler matching is done by the resource-vs-selector check — handler-by-handler.