Configure peering via settings instead of "ourselves"
Previously, the peering setup was stored in a pseudo-peer, so called "ourselves" — as detected at startup. This also included the "legacy peering mode" (which is now removed), and the whole peering configuration — namespace, peering object name, priority, our own identity, etc.

Now, this is split: the peering settings are moved to settings; the dynamic runtime identity is used directly; so as the namespace (with potentially multiple namespaces for multiple peering tasks in mind).

The `Peer` class is only left for the actual peers as read from the peering object, and solely for the purpose of doing the math with lifetime & deadlines.