Make the peering CLI options defaultless
With the default values, they override the default values defined in the settings object. Which is not a big deal until they match, but can be an issue if the configuration defaults are changed later. It is better to have one and only one source of truth for the default values.

A downside: the defaults will not be seen in `--help`. But these values can be misleading anyway, as the operator can override the values in the startup handlers.