Refactor to safe resolution of fields in dicts by default
`assume_empty=True` was enabled for almost all cases where `dicts.resolve()` was used. And in those where it was not enabled, it should be, and the absence of it looked like a mistake.

By this change, simplify the codebase, and use the safe mode always when a default value is specified. This dict resolution is not a generic routine, but is specific to our usage in this framework only, and for K8s, it makes sense to expect the data corruption from outside — instead of failing the process the resource forever after corrupted.