Eliminate python-client from functional tests
The presence/absence of the official Kubernetes client for Python complicates the build matrix and makes it visually clumsy: it is tested in only one or two jobs (which is sufficient), which requires defining the full build matrix, and then excluding its half, and the re-including those one or two jobs manually — to satisfy GitHub Actions' design.

In fact, the official client library is almost not used and completely not tested in the framework: since authentication works with pykube-ng in all testing environments (K3s/K3d, Minikube, etc), the official client is never used alone to be tested.

We can rely on the existing unit-tests where the presence/absence of both clients (pykube-ng and the "officia" one) is simulated with mocks. The tests are rudimentary, but presumably sufficient.

The remaining places where the official client is used can also be removed in favour of more lightweight pykube-ng.