Extract K3d/K3s setup as a reusable GitHub action
The extracted action also contains the latest version detection, which would be an overkill for this repo (unrelated complexity).

As the main benefit, the job names are not semantically limited to K8s `major.minor` versions, but flexible on `.micro` and `+k3s…` versioning. This allows us to configure the required GitHub checks for a PR to be mergeable.