Support resources with no uid, and maybe with no name
This goes slightly against K8s API specification, but the reality is that some resources, e.g. `v1/ComponentStatus`, do not have uids. Thos should not fail the operator.

For such cases, use a fallback scenario, where a sufficiently unique pseudo-uid is generated from those parts of the resource, that do not change over time, and ensure its time-space uniqueness (i.e. if a resource is deleted and a new one is created with the same name, it should be a different pseudo-uid).