Simplify the toggle state changes and checks
In some cases, the toggle should be turned to or checked against a runtime value. With separate methods for on/off values, it requires `if-else` expressions or statements. It could be much more simple if there is only one method to turning it to any state, and one method for waiting, and one method for checking.

Especially in the light of coming changes when the freeze toggles are split to toggles & checkers, and there could be more than one toggle per checker — having too many methods is confusing and prone to errors.

However, the on/off checks are left as is — to keep the code readability, and to avoid confusion when checking for `bool(toggle_set)`, where the "toggle set" is a collection of toggles: a collection is typically true when it contains any items, but we need a special logic over these items.