Reunite the watching coroutines in unfrozen state
In ec2c152c5a069201ed6c40a8306184ca25139198, a simple streaming coroutine has been split to monitor for freeze flags/modes on the streaming itself (before that, the stream was always on, but the events were ignored). Now, that freeze controlling logic requires much more sophistication, so it is extracted into a separate context manager, while the streaming routines are re-united back:

* one for the infinite watch — until the operator exits or the resource becomes out of scope (the latter is not yet done);
* one for the continuous watch: one initial listing + multiple streams — as long as the resource version is valid;
* and one for a single watch (a single API call).

This also simplifies the tests for freezing/resuming of the streams — no mocked streams are needed anymore.