Hide all public interfaces of OperatorRegistry
The users should only use it as an object that is passed to the `registry=` option of the decorators and the `operator()` routines. They should never rely on any methods of accessing the content of the registry either for reading or for writing.

By making these fields protected (underscore-named), we declare that, despite the fields are accessible, their naming, meaning, and own signatures can change any time without warning or ensuring backward compatibility.

As part of this change, also hide those specialised sub-registries classes from the public interface of the framework.

Originally, there was an idea to make a `OperatorRegistryAccessor` or `OperatorRegistryView` wrapper for internal use only. This would solve an issue with accessing the protected fields all over the code. But the severity of the issue is low: it only looks not nice, and might be caught by the linters (questionable) — while the accessor/view class would add complexity.