Run explicit commands in the operator's environment
This mostly includes the authentication vault & login handlers, and Ctrl-C & SIGTERM handlers — but also all future supporting tasks that are running in the background during the normal operator execution (for example, cluster observations and resource/namespace scanning).

Note: the massive diff is only a right shift for the code — with no actual changes inside.