Inject the script as a module for other libraries to detect it
Pydantic was failing while trying to access `sys.modules[...]` with the name of the script file somewhere in its metaclasses.

I'm not sure this is the proper way to inject modules, but it solves the problem, and it is how it is documented by example e.g. here (with the `sys.modules` injection):

* https://docs.python.org/3/library/importlib.html#checking-if-a-module-can-be-imported

The module name is fully faked — not even resembles `pkg.mod` — mostly to simulate the `__main__` pseudo-module; in this case, there can be a few "main" pseudo-modules.