Replace zalando.org with kopf.dev for peering
The peering selectors are modified so that they only select by name (plural name in this case), ignoring the group. As such, both `zalando.org` and `kopf.dev` peerings will be supported (so as any other domains, actually; but the resource name itself implies Kopf, not some other frameworks).

This relies on an assumption that *either* `zalando.org` *or* `kopf.dev` peering resources are installed in the cluster, *not both*. It is assumed that both resources cannot co-exist because they share the same names and would cause a conflict anyway.

*Compatibility*: The newly configured clusters will use `kopf.dev`, and all is fine. The existing cluster can continue using `zalando.org`, and it will work too (unless reconfigured).

*For transitioning*, the old CRD must be removed, and a new CRD must be created. This will wipe all existing peering CRs too, so they must be re-created with the new domain (`kopf.dev`). Applying them with the old domain will fail, as after such a transition, those resources will not exist.