Use Python defaulting for resources, stop simulating it
The mandatory fields were added to make sure that all fields are passed in all places. As a result, it ended up trying to outsmart Python itself by using our own defaults as dicts/kwargs all over the codebase.

Revert this approach back to using Python's features, specifically the default values for data classes. If anything important is really missing at creation of a resource object —which happens only in one place in the whole codebase— it should be tested there explicitly (it is), or detected by breaking some features that rely on that information (which should be tested too).