Call the user-defined callback only if other criteria match
Otherwise, the callback is called even if labels/annotations/fields do not match, and —even worse— when the resource does not match the selector! This is because `all([a, b, c])` calculates all the values initially, while `a and b and c` does the lazy evaluation.

As a result, if 2 different resources A & B are served, each existing handler will be checked against each served resource, and the callbacks intended for a resource A will be checked against the resource B too — despite the resource B does not match the selector of the resource A and is known to be skipped anyway.

This, in turn, can raise unexpected errors if the callbacks assume some fields existence, and fail with e.g. `KeyError` when accessing known fields of a resource A when actually applied to a resource B, which does not have such fields.