Drop pykube-ng dependency (unused)
`pykube-ng` was an underlying client library once, but later replaced with the raw `aiohttp`. The dependency was not removed since then.

Now, if an operator uses `pykube-ng`, it should be declared as a dependency of that operator explicitly, not relied on pre-existing due to Kopf.

For authentication, rudimentary methods (username+password, tokens, client ssl certs) are supported by Kopf directly. More complex authentication schemas (e.g. for GKE with token rotation) should pre-install `pykube-ng` or `kubernetes` explicitly.