Parametrize handlers
Add an optional parameter (`param=` option) to all handlers — for a rudimentary parametrization to distinguish several non-overlapping handlers pointing to the same function. This can be applied to parametrize several causes & filters of the same resource, so as for handlers for different resources (still pointing to the same function, e.g. `children_updated`).