Record exceptions' reprs for throttling errors
In tests, in the `caplog` fixture, the exception stack trace is not converted to strings yet (it is done later in the standard logging). As a result, it is difficult to debug tests when neither the exception nor its stack trace is known at the breakpoints. By this fix, the exceptions' rudimentary info will be put directly to the exception message collected by `caplog`.

This problem does not affect normal logging to stdout/stderr or other streams, which get the fully formatted exception stack traces.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>