Prevent massive toggle creation for non-indexable resources
If the resource is non-indexable, there is no need to create the individual toggles neither for the resource kind nor for each individual resource object — such resources do not block the operator's readiness (but can be blocked themselves).

In practice, even if created, those toggles will be turned on and/or dropped as soon as the processing routine is entered and finds that there are no indexing handlers (nothing to do). This is a waste of computing power and memory on startup, which can affect the operators in massive clusters — for no benefit.

Instead, only create resource kind's & resource object's toggles when and if there are indexing handlers to be executed.

As a particular case, when an operator has no indices at all (assumed to be the default and the most common case), create no single toggle, or at most one (for the global operator block until all specialised per-resource-kind blockers are checked & skipped). Without this fix, there will be as many useless toggles as there are resource kinds (usually below ten) PLUS resource objects (can be in hundreds or thousands).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>