Shift kwargs construction directly to cause classes (for cohesion)
In the historically established design, the kwargs were formed by the invocation routines, for which they had to know each and every cause class and extract its fields as kwargs. As a result, when more and more causes were added over time, the kwarg-building routine was growing.

The worst part of this was that the cause classes could not be moved to their relevant modules where they are locally used: e.g. `DaemonCause` is needed only within `daemons.py` and nowhere else; the same for `ActivityCause` and `activities.py`. Any attempt to move the specialised causes to their specialised modules led to an extra dependency of the "invocation" module on these specialised modules, which in turn often depended on the "invocation" module (directly or indirectly via the "handling" module) — thus creating a circular dependency.

With this split & shift of kwargs construction, the "invocation" module has no need to know about all cause classes existing in the framework, it only needs a source of kwargs — the dynamic nature of kwargs is implemented via polymorphism of cause classes.

Some extra complexity is added for the separation of sync-vs-async — which is currently only used in daemons for the daemon-stopping flags: they must have specialised `stopped.wait()` method. The same for overriding kwargs (now: only for indices). This is an acceptable trade-off for better cohesion and overall code quality.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>