Unquote callbacks as unspecific callables at runtime
In a recent commit, callbacks were properly defined, but this required MyPy extensions (to be released in Python & typing_extensions on Python 3.10 release, presumably Oct'2021). The existence of these extensions required a conditional to avoid installing them at runtime (where they are not needed). Thus, in turn, required quoting (ForwardRef'ing) all references to callbacks.

As a result, this broke IDE navigation, as PyCharm does not understand "quoted" types.

To fix that, a little improvement is added: for runtime, define the same callback protocols, but with no arguments specified — make them just callables. This allows us to unquote the callback types and to use them both at runtime and type-checking time naturally.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>