Accept custom memo types properly
Memos are usually of type `kopf.Memo` but can be configured to any custom type (in embedded operators mostly). As such, the framework must not rely on any behaviour of memo instances, so it uses `object` as the type which has no methods or properties or fields.

However, for the user-facing signatures of handler callbacks, it is declared as `Any` for the same purpose: to accept any specialised custom type. Users are encouraged to put their own classes as annotations of the `memo` kwarg; `kopf.Memo` works as well.

The original declaration of handler callbacks with `memo: AnyMemo` was problematic as it was forcing the operator developers to accept `memo` exactly as this type only: `Union[kopf.Memo, object]`, not just `kopf.Memo` alone — which was against the intended usage scenario with arbitrary custom types or the default type.

For even stricter control, to not mess with any other value by accident (as `object` would accept anything), it is redeclared as a _new type_ "AnyMemo". For the user-facing embedded operator functions, it is defined as literally `object` to prevent exposing the declared new type `AnyMemo`. The type-casting of pure `object` to redeclared `AnyMemo=object` is done at the earliest possible point internally, hidden from the users.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>